- s3d --no-rc does not call kill(kidpid, SIGUSR1) if there is no kid - remove GLUT from autoconf recommendations (but it's still possible to select) - cleanup old INSTALL

+++ server/main.c
@@ -177,7 +177,8 @@ int init()
 	        errn("init():signal()",errno);
     if (signal(SIGTERM, sigint_handler) == SIG_ERR) 
 	        errn("init():signal()",errno);
-	kill(kidpid, SIGUSR1);
+	if (kidpid!=0)
+		kill(kidpid, SIGUSR1);
 #endif
 	return(0);
 }
